requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.4.26
click==8.2.0
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.10
jiter==0.9.0
openai==1.79.0
passlib==1.7.4
Pillow==10.0.1
pyasn1==0.6.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.13.2
uvicorn==0.23.2
